reorg Operations

+++ lib-core/src/main/java/de/mhus/lib/core/operation/NotSuccessful.java
@@ -19,14 +19,14 @@ import de.mhus.lib.core.util.MUri;
 
 public class NotSuccessful extends OperationResult {
 
-    public NotSuccessful(String path, String msg, long rc) {
+    public NotSuccessful(String path, String msg, int rc) {
         setSuccessful(false);
         setMsg(MUri.implodeKeyValues("m", msg));
         setOperationPath(path);
         setReturnCode(rc);
     }
 
-    public NotSuccessful(Operation operation, String msg, long rc) {
+    public NotSuccessful(Operation operation, String msg, int rc) {
         setSuccessful(false);
         setMsg(MUri.implodeKeyValues("m", msg));
         setReturnCode(rc);
@@ -36,7 +36,7 @@ public class NotSuccessful extends OperationResult {
         }
     }
 
-    public NotSuccessful(Operation operation, String msg, String caption, long rc) {
+    public NotSuccessful(Operation operation, String msg, String caption, int rc) {
         setSuccessful(false);
         setMsg(MUri.implodeKeyValues("m", msg, "c", caption));
         setReturnCode(rc);
