+++ lib-core/src/main/java/de/mhus/lib/core/operation/SuccessfulMap.java
@@ -27,11 +27,11 @@ public class SuccessfulMap extends Successful {
         setResult(new MProperties());
     }
 
-    public SuccessfulMap(String path, String msg, long rc) {
+    public SuccessfulMap(String path, String msg, int rc) {
         super(path, msg, rc, new MProperties());
     }
 
-    public SuccessfulMap(String path, String msg, long rc, String... keyValues) {
+    public SuccessfulMap(String path, String msg, int rc, String... keyValues) {
         super(path, msg, rc, keyValues);
     }
